cowtest sbrk failed

+++ kernel/defs.h
@@ -67,7 +67,7 @@ void            kinit(void);
 int             kcount(void);
 void            set_ref_count(uint64, int);
 void            clear_ref_count(uint64);
-int             get_ref_count(uint64);
+uint8             get_ref_count(uint64);
 
 // log.c
 void            initlog(int, struct superblock*);
